backend/requirements.txt
# Strands Agents SDK
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# AWS SDK
boto3>=1.34.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0

